+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ## [Unreleased]
 
+### Added
+
+- [Cable] Add support for `stop_stream_from` and `stop_stream_for`  by [@Digvijay](https://github.com/Digvijay-x1) (#217).
+
 ## [1.21.1] - 2026-02-27
 
 - Improve styling for the `skills` CLI (#223).
@@ -414,4 +418,3 @@
     - support the `root`, `get`, `post`, `patch`, `put`, `delete` methods;
     - support the `scope` method with the `path` and `module` options;
     - support `host` constraint;
-